refactor: private-checklists를 folders 하위로 이동
