+++ server/src/app.module.ts
@@ -1,16 +1,16 @@
 import { Module } from '@nestjs/common';
+import { ConfigModule } from '@nestjs/config';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { ConfigModule } from '@nestjs/config';
 import { CommonModule } from './common/common.module';
-import { UsersModule } from './users/users.module';
-import { UserModel } from './users/entities/user.entity';
-import { FoldersModule } from './folders/folders.module';
 import { FolderModel } from './folders/entities/folder.entity';
-import { ChecklistsModule } from './private-checklists/private-checklists.module';
-import { PrivateChecklistModel } from './private-checklists/entities/private-checklist.entity';
+import { FoldersModule } from './folders/folders.module';
+import { PrivateChecklistModel } from './folders/private-checklists/entities/private-checklist.entity';
+import { ChecklistsModule } from './folders/private-checklists/private-checklists.module';
 import { SharedChecklistModel } from './shared-checklists/entities/shared-checklist.entity';
+import { UserModel } from './users/entities/user.entity';
+import { UsersModule } from './users/users.module';
 
 @Module({
   imports: [
