+++ server/src/folders/entities/folder.entity.ts
@@ -1,7 +1,7 @@
-import { BaseModel } from '../../common/entity/base.entity';
 import { Column, Entity, ManyToOne, OneToMany } from 'typeorm';
-import { PrivateChecklistModel } from '../../private-checklists/entities/private-checklist.entity';
+import { BaseModel } from '../../common/entity/base.entity';
 import { UserModel } from '../../users/entities/user.entity';
+import { PrivateChecklistModel } from '../private-checklists/entities/private-checklist.entity';
 
 @Entity()
 export class FolderModel extends BaseModel {
