+++ server/src/folders/private-checklists/private-checklists.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { FoldersModule } from '../folders/folders.module';
-import { UsersModule } from '../users/users.module';
+import { UsersModule } from '../../users/users.module';
+import { FoldersModule } from '../folders.module';
 import { PrivateChecklistModel } from './entities/private-checklist.entity';
 import { PrivateChecklistsController } from './private-checklists.controller';
 import { PrivateChecklistsService } from './private-checklists.service';
