+++ server/test/test-common.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
 import { ConfigModule } from '@nestjs/config';
+import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserModel } from 'src/users/entities/user.entity';
 import { FolderModel } from '../src/folders/entities/folder.entity';
-import { PrivateChecklistModel } from '../src/private-checklists/entities/private-checklist.entity';
+import { FoldersModule } from '../src/folders/folders.module';
+import { PrivateChecklistModel } from '../src/folders/private-checklists/entities/private-checklist.entity';
 import { SharedChecklistModel } from '../src/shared-checklists/entities/shared-checklist.entity';
 import { UsersModule } from '../src/users/users.module';
-import { FoldersModule } from '../src/folders/folders.module';
 
 @Module({
   imports: [
